Added more info if lapack fails removed integer*8

+++ src/7vci.f90
@@ -110,7 +110,7 @@ CONTAINS
   !enddo
   !end subroutine
 
-  INTEGER*8 FUNCTION snumvecsize()
+  INTEGER FUNCTION snumvecsize()
     USE global
     USE modstates
     IMPLICIT NONE
